feat: Port FlintJVM to Raspberry Pi 4 (AArch64)
- Implement RPi Native HAL (com.rpi.Hardware) with UART, GPIO, Timer, RNG
- Fix 64-bit compatibility: uintptr_t function pointers, pointer casts, array dimensions
- C++14 compatibility: replace inline constexpr with static const, fix copy constructors
- Add embedded class loader to serve .class files from binary (no filesystem needed)
- Create bare-metal Circle kernel that initializes JVM heap and runs HelloRPi.java
- Rebuild libwvm.a and libnative.a for AArch64 RPi4
- Produce kernel8-rpi4.img ready for QEMU

+++ addon/FlintJVM/Native/flint_native_class.cpp
@@ -314,7 +314,7 @@ jobjectArray NativeClass_GetDeclaredMethods0(FNIEnv *env, jclass cls) {
                 etypes->clearProtected();
                 break;
             }
-            method->getField(env->exec, "entry")->setInt32((int32_t)methodInfo);
+            method->getField(env->exec, "entry")->setInt32((intptr_t)methodInfo);
 
             array->getData()[aidx++] = method;
             isOk = true;
@@ -361,7 +361,7 @@ jobjectArray NativeClass_GetDeclaredConstructors0(FNIEnv *env, jclass cls) {
                 etypes->clearProtected();
                 break;
             }
-            ctor->getField(env->exec, "entry")->setInt32((int32_t)methodInfo);
+            ctor->getField(env->exec, "entry")->setInt32((intptr_t)methodInfo);
 
             array->getData()[aidx++] = ctor;
             isOk = true;
